Fix signed int overflow in disappearing timer UI message.

+++ app/src/main/java/org/thoughtcrime/securesms/database/model/GroupsV2UpdateMessageConverter.kt
@@ -384,7 +384,7 @@ object GroupsV2UpdateMessageConverter {
       updates.add(
         GroupChangeChatUpdate.Update(
           groupExpirationTimerUpdate = GroupExpirationTimerUpdate(
-            expiresInMs = change.newTimer!!.duration * 1000,
+            expiresInMs = (change.newTimer!!.duration * 1000L).toUInt().toInt(),
             updaterAci = if (editorUnknown) null else change.editorServiceIdBytes
           )
         )
